requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
Django==2.2.3
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
drf-yasg==1.16.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pep8==1.7.1
Pillow==6.1.0
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.99
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
zope.interface==4.6.0
